requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.1
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
openai>=1.10.0
faiss-cpu>=1.7.4
PyMuPDF>=1.23.14
pypdf>=3.17.4
ollama>=0.1.0
